fix review comments
fix barrier for 1x
